+++ src/ScottPlot5/ScottPlot5/Plottables/SignalXY.cs
@@ -27,8 +27,8 @@ public class SignalXY(ISignalXYSource dataSource) : IPlottable, IHasLine, IHasMa
     /// </summary>
     public ConnectStyle ConnectStyle { get; set; } = ConnectStyle.Straight;
 
-    public int MinRenderIndex { get => Data.MinimumIndex; set => Data.MaximumIndex = value; }
-    public int MaxRenderIndex { get => Data.MinimumIndex; set => Data.MaximumIndex = value; }
+    public int MinRenderIndex { get => Data.MinimumIndex; set => Data.MinimumIndex = value; }
+    public int MaxRenderIndex { get => Data.MaximumIndex; set => Data.MaximumIndex = value; }
 
     public Color Color
     {
@@ -54,6 +54,9 @@ public class SignalXY(ISignalXYSource dataSource) : IPlottable, IHasLine, IHasMa
 
     public virtual void Render(RenderPack rp)
     {
+        if (!IsVisible || Data.Count == 0)
+            return;
+
         Pixel[] markerPixels = Data.GetPixelsToDraw(rp, Axes, ConnectStyle);
 
         Pixel[] linePixels = ConnectStyle switch