+++ CHANGELOG.md
@@ -20,6 +20,14 @@ _Not yet on NuGet..._
 * CoordinateLine: Added additional constructors for creating lines given a point and slope (#3987, #3986) @aalgrou
 * DataLogger: Added `Clear()` and `ResetMinAndMaxValues()` to the data logger source class (#3993, #3969) @jpgarza93
 * Controls: Improved behavior of middle-click-drag zooming over axis panels for plots using DPI scaling (#3994) @bforlgreen
+* Style: Added `Plot.Axes.Hairline()` to enable axis frames, tick marks, and grid lines to render 1px wide regardless of scale factor (#3995) @bforlgreen
+* Axes: Display no ticks instead of throwing an exception if automatic DateTime ticks are used with invalid ranges (#4001) @githubkau
+* SignalXY: Improve support for data sources containing zero-length arrays (#4000) @githubkau
+* CoordinateRect: Added constructor that accepts `IAxes` (#4008, #3985) @CoderPM2011
+* CoordinateRect: Fixed an issue that caused `BottomRight` to return incorrect coordinates (#4009, #3996) @CoderPM2011
+* Cookbook: Added a demonstration of stacked filled line plots (#4010, #3967) @CoderPM2011 @MarkG008
+* Benchmark: Improved text default alignment of double-click benchmark (#4014) @banncan
+* SignalXY: Improved behavior of `MinRenderIndex` and `MaxRenderIndex` (#4011) @StendProg
 
 ## ScottPlot 5.0.35
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-06-10_